Check Recoverable Change Not Working
To get started on https://github.com/imachug/sslcrypto/issues/18, I added a function to check if the public key is not specified, and if it's recoverable, and if so, to use the recovered key as the key and continue as normal. It recovers the key perfectly, but for some reason, it will always mark the signature as being correct, irregardless of whether the data given matches the signature or not.

When the public key is specified, the function acts as normal, and will reject bad data and vice versa. 

The code being used for testing is called check_verify_recoverable.py in the root of the Git repo.